♻️ refactor: 인기 도서 조회 api 오류 수정

See merge request s12-webmobile3-sub1/S12P11A206!76

+++ Backend/src/main/java/com/corp/bookiki/bookhistory/dto/BookRankingResponse.java
@@ -46,6 +46,6 @@ public class BookRankingResponse {
 		description = "도서 대출 횟수",
 		example = "42"
 	)
-	private Long borrowCount;
+	private Integer borrowCount;
 
 }
